refactor(cli): extract diff parser owner
